cdk/requirements.txt
-e .
aws_cdk.core==1.70.0
aws_cdk.aws_s3==1.70.0
aws-cdk.aws_route53==1.70.0
aws-cdk.aws_route53_targets==1.70.0
aws-cdk.aws_certificatemanager==1.70.0
aws_cdk.aws_cloudfront==1.70.0
aws_cdk.aws_cloudwatch==1.70.0
aws_cdk.aws_cloudwatch_actions==1.70.0
aws_cdk.aws_sns==1.70.0
